tests/wpt/web-platform-tests/shadow-dom/elements-and-dom-objects/the-shadow-html-element/test-001.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_10_05_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#shadow-element">
<meta name="assert" content="The shadow HTML element: fallback content">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	var host = d.createElement('div');
	d.body.appendChild(host);

	//Shadow root to play with
	var s = host.createShadowRoot();

	var div = d.createElement('div');
	div.innerHTML = '' +
		'<span id="spandex">This is a shadow root content</span>' +
		'<shadow><span id="shadowId">This is a shadow fallback content</span></shadow>';
	s.appendChild(div);

	assert_true(s.querySelector('#shadowId').offsetTop == 0, 'Shadow fallback content should not be rendered in this case');

}), 'A_10_05_01_T01');

test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	var host = d.createElement('div');
	d.body.appendChild(host);

	//Shadow root to play with
	var s = host.createShadowRoot();

	var div = d.createElement('div');
	div.innerHTML = '' +
		'<span id="spandex">This is a shadow root content</span>' +
		'<content><span id="contentId">This is a content fallback</span></content>';
	s.appendChild(div);

	assert_true(s.querySelector('#contentId').offsetTop > 0, 'Fallback content should be rendered');

}), 'A_10_05_01_T02');
</script>
</body>
</html>
